Removed CHANGELOG and simplified Debian release workflow
Previously CHANGELOG had to be updated manually. This was done by downloading
the output of a CI job that used listed commits since the last release. This
made releasing a bit more complicated and led to uselsess release commits.

Now a script creates the full debian/changelog with all versions based on
git tags and git commits.